rename package fix detekt issues
