+++ core/remote-config/src/commonMain/kotlin/xyz/ksharma/krail/core/remoteconfig/di/RemoteConfigModule.kt
@@ -1,13 +1,13 @@
-package xyz.ksharma.krail.core.remote_config.di
+package xyz.ksharma.krail.core.remoteconfig.di
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
 import org.koin.dsl.module
-import xyz.ksharma.krail.core.remote_config.RealRemoteConfig
-import xyz.ksharma.krail.core.remote_config.RemoteConfig
-import xyz.ksharma.krail.core.remote_config.flag.Flag
-import xyz.ksharma.krail.core.remote_config.flag.RemoteConfigFlag
+import xyz.ksharma.krail.core.remoteconfig.RealRemoteConfig
+import xyz.ksharma.krail.core.remoteconfig.RemoteConfig
+import xyz.ksharma.krail.core.remoteconfig.flag.Flag
+import xyz.ksharma.krail.core.remoteconfig.flag.RemoteConfigFlag
 
 val remoteConfigModule = module {
     single<RemoteConfig> {
