+++ core/remote-config/src/commonMain/kotlin/xyz/ksharma/krail/core/remoteconfig/flag/FlagKeys.kt
@@ -1,4 +1,4 @@
-package xyz.ksharma.krail.core.remote_config.flag
+package xyz.ksharma.krail.core.remoteconfig.flag
 
 enum class FlagKeys(val key: String) {
     HIGH_PRIORITY_STOP_IDS("high_priority_stop_ids"),
