+++ core/test/src/commonTest/kotlin/xyz/ksharma/core/test/fakes/FakeFlag.kt
@@ -1,8 +1,8 @@
 package xyz.ksharma.core.test.fakes
 
-import xyz.ksharma.krail.core.remote_config.flag.Flag
-import xyz.ksharma.krail.core.remote_config.flag.FlagKeys
-import xyz.ksharma.krail.core.remote_config.flag.FlagValue
+import xyz.ksharma.krail.core.remoteconfig.flag.Flag
+import xyz.ksharma.krail.core.remoteconfig.flag.FlagKeys
+import xyz.ksharma.krail.core.remoteconfig.flag.FlagValue
 
 class FakeFlag : Flag {
     private val flagValues = mutableMapOf<String, FlagValue>()
