+++ feature/trip-planner/ui/src/commonMain/kotlin/xyz/ksharma/krail/trip/planner/ui/savedtrips/SavedTripsViewModel.kt
@@ -26,14 +26,13 @@ import kotlinx.datetime.toLocalDateTime
 import xyz.ksharma.krail.core.analytics.Analytics
 import xyz.ksharma.krail.core.analytics.AnalyticsScreen
 import xyz.ksharma.krail.core.analytics.event.AnalyticsEvent
-import xyz.ksharma.krail.core.analytics.event.AnalyticsEvent.*
 import xyz.ksharma.krail.core.analytics.event.trackScreenViewEvent
 import xyz.ksharma.krail.core.log.log
 import xyz.ksharma.krail.core.log.logError
-import xyz.ksharma.krail.core.remote_config.flag.Flag
-import xyz.ksharma.krail.core.remote_config.flag.FlagKeys
-import xyz.ksharma.krail.core.remote_config.flag.asBoolean
-import xyz.ksharma.krail.core.remote_config.flag.asNumber
+import xyz.ksharma.krail.core.remoteconfig.flag.Flag
+import xyz.ksharma.krail.core.remoteconfig.flag.FlagKeys
+import xyz.ksharma.krail.core.remoteconfig.flag.asBoolean
+import xyz.ksharma.krail.core.remoteconfig.flag.asNumber
 import xyz.ksharma.krail.coroutines.ext.launchWithExceptionHandler
 import xyz.ksharma.krail.info.tile.network.api.InfoTileManager
 import xyz.ksharma.krail.info.tile.state.InfoTileData
@@ -625,7 +624,6 @@ class SavedTripsViewModel(
     }
 
     private fun onInfoTileCtaClick(infoTile: InfoTileData) {
-
         when (infoTile.type) {
             InfoTileData.InfoTileType.INVITE_FRIENDS -> {
                 platformOps.sharePlainText(
@@ -633,15 +631,16 @@ class SavedTripsViewModel(
                     title = "Invite your Friends",
                 )
                 analytics.track(
-                    event = ReferFriend(
-                        entryPoint = ReferFriend.EntryPoint.SAVED_TRIPS,
+                    event = AnalyticsEvent.ReferFriend(
+                        entryPoint = AnalyticsEvent.ReferFriend.EntryPoint.SAVED_TRIPS,
                     ),
                 )
             }
 
             InfoTileData.InfoTileType.CRITICAL_ALERT,
             InfoTileData.InfoTileType.INFO,
-            InfoTileData.InfoTileType.APP_UPDATE -> {
+            InfoTileData.InfoTileType.APP_UPDATE,
+            -> {
                 infoTile.primaryCta?.url?.let { url ->
                     platformOps.openUrl(url)
                     trackInfoTileInteraction(
